% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos_files.R
\name{argos}
\alias{argos}
\title{Automatic Regression for Governing Equations (ARGOS)}
\usage{
argos(
  sorted_theta,
  xdot,
  monomial_orders,
  library_type = c("poly", "four", "poly_four"),
  state_var_deriv = 1,
  alpha_level = 0.05,
  num_samples = 2000,
  sr_method = c("lasso", "alasso"),
  weights_method = NULL,
  ols_ps = TRUE,
  parallel = c("no", "multicore", "snow"),
  ncpus = NULL
)
}
\arguments{
\item{sorted_theta}{A matrix or data frame containing sorted values of the variable 'theta'.}

\item{xdot}{A matrix or data frame containing derivative data.}

\item{monomial_orders}{A vector indicating the monomial order for each term in the `sorted_theta`.}

\item{library_type}{A character vector (default: c("poly", "four", "poly_four")) specifying the type of library being used.}

\item{state_var_deriv}{An integer. The index of the state variable for which the derivative is calculated. Default is 1.}

\item{alpha_level}{A numeric scalar. The level of significance for confidence intervals. Default is 0.05.}

\item{num_samples}{An integer. The number of bootstrap samples. Default is 2000.}

\item{sr_method}{A character string. The sparse regression method to be used, either "lasso" or "alasso". Default is "lasso".}

\item{weights_method}{A string or NULL. The method for calculating weights in the Adaptive Lasso. If NULL, ridge regression pilot estimates are used. Default is NULL.}

\item{ols_ps}{A logical. If TRUE, post-selection OLS is performed after the Lasso or Adaptive Lasso. Default is TRUE.}

\item{parallel}{A character string. The type of parallel computation to be used, either "no", "multicore" or "snow". Default is "no".}

\item{ncpus}{An integer or NULL. The number of cores to be used in parallel computation. If NULL, the function will try to detect the number of cores. Default is NULL.}
}
\value{
A list with three elements:
  - point_estimates: a vector of point estimates for the coefficients.
  - ci: a matrix where each column represents the lower and upper bounds of the confidence interval for a coefficient.
  - identified_model: a matrix of coefficients of the identified model.
}
\description{
This function performs sparse regression on a dataset to identify the governing equations
of the system. It uses the Savitzky-Golay filter to preprocess the data and then applies
the Lasso or Adaptive Lasso for feature selection.
}
