% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos_files.R
\name{build_design_matrix}
\alias{build_design_matrix}
\title{Build Design Matrix}
\usage{
build_design_matrix(
  x_t,
  dt = 1,
  sg_poly_order = 4,
  library_degree = 5,
  library_type = c("poly", "four", "poly_four")
)
}
\arguments{
\item{x_t}{Matrix of observations.}

\item{dt}{Time step (default is 1).}

\item{sg_poly_order}{Polynomial order for Savitzky-Golay Filter.}

\item{library_degree}{Degree of polynomial library (default is 5).}

\item{library_type}{Type of library to use. Can be one of "poly", "four", or "poly_four"}
}
\value{
A list with two elements:
\itemize{
  \item \code{sorted_theta} - A matrix with sorted polynomial/trigonometric terms.
  \item \code{monomial_orders} - A vector indicating the order of each polynomial term.
  \item \code{xdot_filtered} - A matrix with derivative terms (dependent variable).
}
}
\description{
This function first smooths the data and approximates the derivative before building the design matrix to include monomial and fourier terms.
}
