% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamical_systems.R
\name{duffing_oscillator}
\alias{duffing_oscillator}
\title{Simulate a Duffing Oscillator}
\usage{
duffing_oscillator(n_obs, dt, init_conditions, snr)
}
\arguments{
\item{n_obs}{A numeric value specifying the number of observations or time points for which you want to simulate the system.}

\item{dt}{A numeric value specifying the time step size for the simulation.}

\item{init_conditions}{A numeric vector specifying the initial conditions for the simulation. It should contain two elements corresponding to the initial values of the two state variables.}

\item{snr}{A numeric value specifying the signal-to-noise ratio in dB. This is used to add noise to the simulation. If \code{snr} equals 0, no noise is added.}
}
\value{
A matrix where each row corresponds to a time point and each column corresponds to a state variable. The matrix represents the simulated state of the system over time.
}
\description{
This function simulates a duffing system. It is a wrapper for the Python function \code{duffing_ode}, which uses a set of ordinary differential equations (ODEs) to simulate the system. Noise can be added to the system.
}
\examples{
# Simulate a system for 100 time points with a time step size of 0.01,
# initial conditions of c(1, 0) and a signal-to-noise ratio of 49 dB.
sim <- duffing_oscillator(n_obs=5000, 0.01, c(1, 0), 49)
}
